+++ R/as_tibble.R
@@ -1,14 +1,10 @@
 #' Convert gtsummary object to a tibble
 #'
-#' Function converts gtsummary objects tibbles. The formatting stored in
-#' `x$kable_calls` is applied.
+#' Function converts a gtsummary object to a tibble.
 #'
 #' @inheritParams as_kable
 #' @param col_labels Logical argument adding column labels to output tibble.
 #' Default is `TRUE`.
-#' @param group_header If a group is present (e.g. via `tbl_merge(group_header=)`),
-#' this string specifies the column header for the grouping column. Default
-#' is `"**Group**"`
 #' @param ... Not used
 #' @return a [tibble][tibble::tibble-package]
 #' @family gtsummary output types
@@ -25,8 +21,7 @@
 #' # without column labels
 #' as_tibble(tbl, col_labels = FALSE)
 as_tibble.gtsummary <- function(x, include = everything(), col_labels = TRUE,
-                                return_calls = FALSE, group_header = NULL,
-                                exclude = NULL,  ...) {
+                                return_calls = FALSE, exclude = NULL,  ...) {
   # DEPRECATION notes ----------------------------------------------------------
   if (!rlang::quo_is_null(rlang::enquo(exclude))) {
     lifecycle::deprecate_warn(
@@ -41,14 +36,8 @@ as_tibble.gtsummary <- function(x, include = everything(), col_labels = TRUE,
     )
   }
 
-  # setting defaults -----------------------------------------------------------
-  group_header <-
-    group_header %||%
-    get_theme_element("pkgwide-str:group_header", default = "**Group**")
-
   # creating list of calls to get formatted tibble -----------------------------
-  tibble_calls <- table_header_to_tibble_calls(x = x, col_labels = col_labels,
-                                               group_header = group_header)
+  tibble_calls <- table_header_to_tibble_calls(x = x, col_labels = col_labels)
 
   # converting to character vector ---------------------------------------------
   include <- var_input_to_string(data = vctr_2_tibble(names(tibble_calls)),
@@ -79,7 +68,7 @@ as_tibble.gtsummary <- function(x, include = everything(), col_labels = TRUE,
 }
 
 
-table_header_to_tibble_calls <- function(x, col_labels =  TRUE, group_header) {
+table_header_to_tibble_calls <- function(x, col_labels =  TRUE) {
   table_header <- x$table_header
   tibble_calls <- list()
 
@@ -89,13 +78,13 @@ table_header_to_tibble_calls <- function(x, col_labels =  TRUE, group_header) {
   # ungroup --------------------------------------------------------------------
   group_var <- select(x$table_body, dplyr::group_cols()) %>% names()
   if (length(group_var) > 0) {
+    if (group_var != "groupname_col")
+      stop("`.$table_body` may only be grouped by column 'groupname_col'")
+
     tibble_calls[["ungroup"]] <- list(
-      expr(ungroup()),
-      expr(mutate(...new_group_var... = !!sym(group_var))),
-      expr(group_by(.data$...new_group_var...)),
-      expr(mutate_at(vars(!!sym(group_var)), ~ifelse(dplyr::row_number() == 1, ., NA))),
-      expr(ungroup()),
-      expr(select(-.data$...new_group_var...))
+      expr(mutate(groupname_col =
+                    ifelse(dplyr::row_number() == 1, .data$groupname_col, NA))),
+      expr(ungroup())
     )
   }
 
@@ -144,7 +133,7 @@ table_header_to_tibble_calls <- function(x, col_labels =  TRUE, group_header) {
   )
 
   # converting all cols to character...
-  # this is important for some output types, e.g. as_flextable, so missing don't
+  # this is important for some output types, e.g. as_flex_table, so missing don't
   # display as NA
   cols_to_keep <-
     dplyr::filter(table_header, .data$hide == FALSE) %>%
@@ -162,13 +151,8 @@ table_header_to_tibble_calls <- function(x, col_labels =  TRUE, group_header) {
     df_col_labels <-
       dplyr::filter(table_header, .data$hide == FALSE)
 
-    # if there is a grouping variable, add the column header
-    if (length(group_var) > 0)
-      tibble_calls[["cols_label"]] <-
-        expr(rlang::set_names(c(!!group_header, !!df_col_labels$label)))
-    else
-      tibble_calls[["cols_label"]] <-
-        expr(rlang::set_names(!!df_col_labels$label))
+    tibble_calls[["cols_label"]] <-
+      expr(rlang::set_names(!!df_col_labels$label))
   }
 
   tibble_calls