+++ R/print.R
@@ -21,7 +21,7 @@ print.gtsummary <- function(x, print_engine = NULL, ...) {
 
   # checking engine
   accepted_print_engines <-
-    c("gt", "kable", "kable_extra", "flextable", "tibble")
+    c("gt", "kable", "kable_extra", "flextable", "huxtable", "tibble")
   if (!rlang::is_string(print_engine) || !print_engine %in% accepted_print_engines) {
     stop(glue(
       "Select a valid print engine. ",
@@ -34,9 +34,9 @@ print.gtsummary <- function(x, print_engine = NULL, ...) {
     print_engine,
     "gt" = as_gt(x),
     "kable" = as_kable(x),
-    "flextable" = as_flextable(x),
+    "flextable" = as_flex_table(x),
     "kable_extra" = as_kable_extra(x),
-    # "huxtable" = as_huxtable(x),
+    "huxtable" = as_hux_table(x),
     "tibble" = as_tibble(x)
   ) %>%
     print()
@@ -70,13 +70,23 @@ knit_print.gtsummary <- function(x, ...) {
   # don't use word_document with gt engine
   else if (identical(print_engine %||% "gt", "gt") &&
            "docx" %in% knitr::opts_knit$get('rmarkdown.pandoc.to')) {
-    rlang::inform(paste(
-      "Table printed with `knitr::kable()`, not {gt}. Learn why at",
-      "http://www.danieldsjoberg.com/gtsummary/articles/rmarkdown.html",
-      "To suppress this message, include `message = FALSE` in the code chunk header.",
-      sep = "\n"
-    ))
-    print_engine <- "kable"
+    if (requireNamespace("flextable", quietly = TRUE)) {
+      rlang::inform(paste(
+        "Table printed with {flextable}, not {gt}. Learn why at",
+        "http://www.danieldsjoberg.com/gtsummary/articles/rmarkdown.html",
+        "To suppress this message, include `message = FALSE` in the code chunk header.",
+        sep = "\n"
+      ))
+      print_engine <- "flextable"    }
+    else {
+      rlang::inform(paste(
+        "Table printed with `knitr::kable()`, not {gt}. Learn why at",
+        "http://www.danieldsjoberg.com/gtsummary/articles/rmarkdown.html",
+        "To suppress this message, include `message = FALSE` in the code chunk header.",
+        sep = "\n"
+      ))
+      print_engine <- "kable"
+    }
   }
 
   # RTF warning when using gt
@@ -99,9 +109,9 @@ knit_print.gtsummary <- function(x, ...) {
     print_engine,
     "gt" = as_gt(x),
     "kable" = as_kable(x),
-    "flextable" = as_flextable(x),
+    "flextable" = as_flex_table(x),
     "kable_extra" = as_kable_extra(x),
-    # "huxtable" = as_huxtable(x),
+    "huxtable" = as_hux_table(x),
     "tibble" = as_tibble(x)
   ) %>%
       knitr::knit_print()